# docs/openapi/scforge-public.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "SCForge \u5F00\u653E API",
    "description": "SCForge\uFF08\u751F\u5B58\u6218\u4E89\u63D2\u4EF6 / \u6A21\u7EC4\u8D44\u6E90\u5E73\u53F0\uFF09\u7684\u5BF9\u5916\u5F00\u653E\u63A5\u53E3\u3002\n\n**\u9274\u6743**\uFF1A\u8BFB\u63A5\u53E3\u533F\u540D\u53EF\u7528\uFF1B\u4E0E\u300C\u6211\u7684\u8D44\u6E90\u300D\u76F8\u5173\u7684\u8BFB\u5199\u9700\u8981 API Key \u2014\u2014 \u8BF7\u6C42\u5934 Authorization: Bearer scf_\u003Cbase64url\u003E\u3002\u7B7E\u53D1\u4E0E\u4F5C\u7528\u57DF\u89C1\u63A5\u5165\u6307\u5357\u3002\n\n**\u9519\u8BEF\u5F62\u72B6**\uFF1A\u6240\u6709\u9519\u8BEF\u54CD\u5E94\u7EDF\u4E00\u4E3A { \u0022detail\u0022: \u0022\u4E2D\u6587\u8BF4\u660E\u0022 }\uFF08\u89C1 ApiError.cs\uFF09\uFF0CHTTP \u72B6\u6001\u7801\u5373\u8BED\u4E49\uFF08400 \u53C2\u6570\u6216\u89C4\u5219\u4E0D\u5408\u6CD5\u3001401 \u672A\u767B\u5F55\u3001403 \u4F5C\u7528\u57DF\u4E0D\u8DB3\u3001404 \u8D44\u6E90\u4E0D\u5B58\u5728\u3001409 \u72B6\u6001\u51B2\u7A81\uFF09\u3002\n\n**\u65F6\u95F4**\uFF1A\u6240\u6709\u65F6\u95F4\u5B57\u6BB5\u6309\u5317\u4EAC\u65F6\u95F4\uFF08UTC\u002B8\uFF0C\u002B08:00\uFF09\u8F93\u51FA\u3002",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://api.cldery.com",
      "description": "\u751F\u4EA7\u73AF\u5883"
    }
  ],
  "paths": {
    "/scforge/addons": {
      "get": {
        "tags": [
          "ScforgeAddons"
        ],
        "summary": "\u5206\u9875\u641C\u7D22\uFF1Aq / category / tag / gameVersion / sort / kind / page / pageSize\u3002kind \u7701\u7565\u5373\u63D2\u4EF6\u4E0E\u6A21\u7EC4\u90FD\u8FD4\u56DE\u3002",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "category",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tag",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "gameVersion",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "kind",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 24
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "ScforgeAddons"
        ],
        "summary": "\u53D1\u5E03\u65B0\u8D44\u6E90\uFF08multipart/form-data\uFF09\u3002\u63D0\u4EA4\u540E\u8FDB\u5165\u5F85\u5BA1\u6838\uFF0C\u7BA1\u7406\u5458\u901A\u8FC7\u540E\u624D\u5BF9\u516C\u4F17\u53EF\u89C1\u3002\r\nkind \u51B3\u5B9A\u662F\u63D2\u4EF6\uFF08plugin\uFF09\u8FD8\u662F\u6A21\u7EC4\uFF08mod\uFF09\uFF0C\u521B\u5EFA\u540E\u4E0D\u53EF\u66F4\u6539\uFF0C\u5305\u6587\u4EF6\u6309 kind \u6821\u9A8C\u6269\u5C55\u540D\u3002\r\n\u8BF7\u6C42\u4F53\u4E0A\u9650\u6309\u300C\u8D44\u6E90\u5305\u4E0A\u9650 \u002B \u56FE\u7247\u9884\u7B97\u300D\u653E\u5BBD\uFF0C\u5426\u5219\u5927\u5305\u4F1A\u88AB Kestrel \u7684 30MB \u9ED8\u8BA4\u503C\u6321\u4E0B\u3002",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "Package": {
                    "type": "string",
                    "description": "\u8D44\u6E90\u5305\uFF08\u5FC5\u586B\uFF09\uFF1A.dll\uFF08\u63D2\u4EF6\uFF09\u6216 .netmod\uFF08\u6A21\u7EC4\uFF0C\u4F1A\u4E0B\u53D1\u5230\u5BA2\u6237\u7AEF\uFF09\u3002",
                    "format": "binary"
                  },
                  "Kind": {
                    "type": "string",
                    "description": "\u8D44\u6E90\u7C7B\u578B\uFF1Aplugin\uFF08\u63D2\u4EF6\uFF0C\u4EC5\u670D\u52A1\u7AEF\uFF09\u6216 mod\uFF08\u6A21\u7EC4\uFF0C\u4F1A\u4E0B\u53D1\u5230\u5BA2\u6237\u7AEF\uFF09\u3002\u521B\u5EFA\u540E\u4E0D\u53EF\u66F4\u6539\u3002"
                  },
                  "Name": {
                    "type": "string"
                  },
                  "Slug": {
                    "type": "string"
                  },
                  "Summary": {
                    "type": "string"
                  },
                  "Description": {
                    "type": "string"
                  },
                  "Readme": {
                    "type": "string"
                  },
                  "Category": {
                    "type": "string"
                  },
                  "GameVersion": {
                    "type": "string"
                  },
                  "Tags": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "description": "\u6807\u7B7E\u952E\uFF1B\u540C\u540D\u591A\u503C\u53EF\u5728\u8868\u5355\u91CC\u91CD\u590D\u51FA\u73B0\u3002"
                  },
                  "SourceUrl": {
                    "type": "string"
                  },
                  "IssuesUrl": {
                    "type": "string"
                  },
                  "License": {
                    "type": "string"
                  },
                  "LicenseUrl": {
                    "type": "string"
                  },
                  "DonationUrl": {
                    "type": "string"
                  },
                  "DiscordUrl": {
                    "type": "string"
                  },
                  "Icon": {
                    "type": "string",
                    "description": "\u8D44\u6E90\u56FE\u6807\uFF08\u53EF\u9009\uFF09\uFF1AJPG / PNG / WebP / GIF\u3002",
                    "format": "binary"
                  },
                  "Gallery": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    },
                    "description": "\u622A\u56FE\uFF08\u53EF\u9009\uFF0C\u6700\u591A 6 \u5F20\uFF09\u3002"
                  },
                  "Version": {
                    "type": "string"
                  },
                  "Channel": {
                    "type": "string"
                  },
                  "Changelog": {
                    "type": "string"
                  },
                  "GameVersions": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "Dependencies": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                }
              },
              "encoding": {
                "Package": {
                  "style": "form"
                },
                "Kind": {
                  "style": "form"
                },
                "Name": {
                  "style": "form"
                },
                "Slug": {
                  "style": "form"
                },
                "Summary": {
                  "style": "form"
                },
                "Description": {
                  "style": "form"
                },
                "Readme": {
                  "style": "form"
                },
                "Category": {
                  "style": "form"
                },
                "GameVersion": {
                  "style": "form"
                },
                "Tags": {
                  "style": "form"
                },
                "SourceUrl": {
                  "style": "form"
                },
                "IssuesUrl": {
                  "style": "form"
                },
                "License": {
                  "style": "form"
                },
                "LicenseUrl": {
                  "style": "form"
                },
                "DonationUrl": {
                  "style": "form"
                },
                "DiscordUrl": {
                  "style": "form"
                },
                "Icon": {
                  "style": "form"
                },
                "Gallery": {
                  "style": "form"
                },
                "Version": {
                  "style": "form"
                },
                "Channel": {
                  "style": "form"
                },
                "Changelog": {
                  "style": "form"
                },
                "GameVersions": {
                  "style": "form"
                },
                "Dependencies": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/scforge/addons/{idOrSlug}": {
      "get": {
        "tags": [
          "ScforgeAddons"
        ],
        "summary": "\u8BE6\u60C5\uFF1AidOrSlug \u63A5\u53D7 GUID \u6216 slug\u3002",
        "parameters": [
          {
            "name": "idOrSlug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/scforge/addons/{id}": {
      "put": {
        "tags": [
          "ScforgeAddons"
        ],
        "summary": "\u7F16\u8F91\u8D44\u6E90\u8D44\u6599\uFF08\u4F5C\u8005\uFF0Cmultipart\uFF09\uFF1A\u7F16\u8F91\u4F1A\u91CD\u65B0\u8FDB\u5165\u5F85\u5BA1\u6838\u3002",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "Summary": {
                    "type": "string"
                  },
                  "Description": {
                    "type": "string"
                  },
                  "Readme": {
                    "type": "string"
                  },
                  "Category": {
                    "type": "string"
                  },
                  "GameVersion": {
                    "type": "string"
                  },
                  "Tags": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "SourceUrl": {
                    "type": "string"
                  },
                  "IssuesUrl": {
                    "type": "string"
                  },
                  "License": {
                    "type": "string"
                  },
                  "LicenseUrl": {
                    "type": "string"
                  },
                  "DonationUrl": {
                    "type": "string"
                  },
                  "DiscordUrl": {
                    "type": "string"
                  },
                  "Icon": {
                    "type": "string",
                    "description": "\u4F20\u5165\u5373\u66FF\u6362\u56FE\u6807\u3002",
                    "format": "binary"
                  },
                  "Gallery": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    },
                    "description": "\u4F20\u5165\u5373\u6574\u4F53\u66FF\u6362\u622A\u56FE\u5217\u8868\uFF08\u6700\u591A 6 \u5F20\uFF09\u3002"
                  },
                  "ClearIcon": {
                    "type": "boolean",
                    "description": "\u6E05\u7A7A\u56FE\u6807\u3002"
                  },
                  "ClearGallery": {
                    "type": "boolean",
                    "description": "\u6E05\u7A7A\u622A\u56FE\u3002"
                  }
                }
              },
              "encoding": {
                "Summary": {
                  "style": "form"
                },
                "Description": {
                  "style": "form"
                },
                "Readme": {
                  "style": "form"
                },
                "Category": {
                  "style": "form"
                },
                "GameVersion": {
                  "style": "form"
                },
                "Tags": {
                  "style": "form"
                },
                "SourceUrl": {
                  "style": "form"
                },
                "IssuesUrl": {
                  "style": "form"
                },
                "License": {
                  "style": "form"
                },
                "LicenseUrl": {
                  "style": "form"
                },
                "DonationUrl": {
                  "style": "form"
                },
                "DiscordUrl": {
                  "style": "form"
                },
                "Icon": {
                  "style": "form"
                },
                "Gallery": {
                  "style": "form"
                },
                "ClearIcon": {
                  "style": "form"
                },
                "ClearGallery": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "ScforgeAddons"
        ],
        "summary": "\u5220\u9664\u8D44\u6E90\uFF08\u8FDE\u540C\u7248\u672C\u3001\u8BC4\u8BBA\u4E0E\u6295\u7968\uFF09\u3002",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "ScforgeApiKey": {
        "type": "http",
        "description": "SCForge API Key\uFF08scf_ \u5F00\u5934\uFF09\u3002\u8BFB\u63A5\u53E3\u533F\u540D\u4E5F\u53EF\u8C03\u7528\uFF0C\u5E26\u4E0A Key \u624D\u80FD\u8BBF\u95EE\u300C\u6211\u7684\u8D44\u6E90\u300D\u4E0E\u5199\u63A5\u53E3\u3002",
        "scheme": "bearer",
        "bearerFormat": "scf_\u003Cbase64url\u003E"
      }
    }
  },
  "security": [
    {
      "ScforgeApiKey": []
    }
  ],
...
}